[aztfy enhancement proposal] `-o` option for specified output dir

+++ internal/meta/meta_impl.go
@@ -5,6 +5,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"io"
 	"os"
 	"path"
 	"path/filepath"
@@ -32,7 +33,7 @@ type MetaImpl struct {
 	armTemplate    armtemplate.Template
 }
 
-func newMetaImpl(rg string) (Meta, error) {
+func newMetaImpl(rg string, outputDir string) (Meta, error) {
 	ctx := context.TODO()
 
 	// Initialize the workspace
@@ -53,11 +54,35 @@ func newMetaImpl(rg string) (Meta, error) {
 	}
 
 	wsp := filepath.Join(rootDir, rg)
-	if err := os.RemoveAll(wsp); err != nil {
-		return nil, fmt.Errorf("removing existing workspace %q: %w", wsp, err)
-	}
-	if err := os.MkdirAll(wsp, 0755); err != nil {
-		return nil, fmt.Errorf("creating workspace %q: %w", wsp, err)
+
+	if outputDir != "" {
+		wsp = outputDir
+
+		// Ensure wsp is an empty directory
+		stat, err := os.Stat(wsp)
+		if os.IsNotExist(err) {
+			return nil, fmt.Errorf("the output directory %q doesn't exist", wsp)
+		}
+		if !stat.IsDir() {
+			return nil, fmt.Errorf("the output path %q is not a directory", wsp)
+		}
+
+		f, err := os.Open(wsp)
+		if err != nil {
+			return nil, err
+		}
+		_, err = f.Readdirnames(1) // Or f.Readdir(1)
+		f.Close()
+		if err != io.EOF {
+			return nil, fmt.Errorf("the output directory %q is not empty", wsp)
+		}
+	} else {
+		if err := os.RemoveAll(wsp); err != nil {
+			return nil, fmt.Errorf("removing existing workspace %q: %w", wsp, err)
+		}
+		if err := os.MkdirAll(wsp, 0755); err != nil {
+			return nil, fmt.Errorf("creating workspace %q: %w", wsp, err)
+		}
 	}
 
 	// Authentication
